refactor: credential manifest enhancements to match wallet-rendering specs Signed-off-by: heidihan0000 <daeun.han@avast.com>
