+++ pkg/doc/cm/credentialmanifest.go
@@ -13,13 +13,11 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"strings"
 
 	"github.com/PaesslerAG/gval"
 	"github.com/PaesslerAG/jsonpath"
 
 	"github.com/hyperledger/aries-framework-go/pkg/doc/presexch"
-	"github.com/hyperledger/aries-framework-go/pkg/doc/verifiable"
 )
 
 // CredentialManifestAttachmentFormat defines the format type of Credential Manifest when used as an attachment in the
@@ -87,8 +85,9 @@ type DataDisplayDescriptor struct {
 // DisplayMappingObject represents a Display Mapping Object as defined in
 // https://identity.foundation/credential-manifest/wallet-rendering/#display-mapping-object
 // There are two possibilities here:
-// If the text field is used, then schema is not needed.
-// If the path field is used, then schema is required.
+// 1. If the text field is used, schema is not required. The text field will contain display
+// information about the target Claim.
+// 2. If the path field is used, schema is required. Data will be pulled from the target Claim using the path.
 // TODO (#3045) Support for JSONPath bracket notation.
 type DisplayMappingObject struct {
 	Text     string   `json:"text,omitempty"`
@@ -107,56 +106,16 @@ type LabeledDisplayMappingObject struct {
 // Schema represents Type and (optional) Format information for a DisplayMappingObject that uses the Paths field,
 // as defined in https://identity.foundation/credential-manifest/wallet-rendering/#using-path.
 type Schema struct {
-	Type   string `json:"type,omitempty"`   // MUST be here
-	Format string `json:"format,omitempty"` // MAY be here if the Type is "string".
-}
-
-// ResolvedProperty contains resolved result for each resolved property.
-type ResolvedProperty struct {
-	Schema Schema      `json:"schema,omitempty"`
-	Label  string      `json:"label,omitempty"`
-	Value  interface{} `json:"value,omitempty"`
-}
-
-// ResolvedDescriptor typically represents results of resolving manifests by credential fulfillment.
-// typically represents a DataDisplayDescriptor that's had its various "template" fields resolved
-// into concrete values based on a Verifiable Credential.
-type ResolvedDescriptor struct {
-	DescriptorID string              `json:"descriptor_id,omitempty"`
-	Title        string              `json:"title,omitempty"`
-	Subtitle     string              `json:"subtitle,omitempty"`
-	Description  string              `json:"description,omitempty"`
-	Styles       Styles              `json:"styles,omitempty"`
-	Properties   []*ResolvedProperty `json:"properties,omitempty"`
+	Type             string `json:"type,omitempty"`      // MUST be here
+	Format           string `json:"format,omitempty"`    // MAY be here if the Type is "string".
+	ContentMediaType string `json:"mediatype,omitempty"` // MAY be here if the Type is "string".
+	ContentEncoding  string `json:"encoding,omitempty"`  // MAY be here if the Type is "string".
 }
 
 type staticDisplayMappingObjects struct {
 	title       string
 	subtitle    string
 	description string
-}
-
-// resolveCredOpts contains option to provide credential to resolve manifest.
-type resolveCredOpts struct {
-	credential    *verifiable.Credential
-	rawCredential json.RawMessage
-}
-
-// CredentialToResolveOption is option to provide credential to resolve manifest.
-type CredentialToResolveOption func(opts *resolveCredOpts)
-
-// CredentialToResolve to provide verifiable credential instance to resolve.
-func CredentialToResolve(credential *verifiable.Credential) CredentialToResolveOption {
-	return func(opts *resolveCredOpts) {
-		opts.credential = credential
-	}
-}
-
-// RawCredentialToResolve to provide raw JSON bytes of verifiable credential to resolve.
-func RawCredentialToResolve(raw json.RawMessage) CredentialToResolveOption {
-	return func(opts *resolveCredOpts) {
-		opts.rawCredential = raw
-	}
 }
 
 // UnmarshalJSON is the custom unmarshal function gets called automatically when the standard json.Unmarshal is called.
@@ -167,7 +126,7 @@ func (cm *CredentialManifest) UnmarshalJSON(data []byte) error {
 		return err
 	}
 
-	err = cm.Validate()
+	err = cm.validate()
 	if err != nil {
 		return fmt.Errorf("invalid credential manifest: %w", err)
 	}
@@ -177,7 +136,7 @@ func (cm *CredentialManifest) UnmarshalJSON(data []byte) error {
 
 // Validate ensures that this CredentialManifest is valid as per the spec.
 // Note that this function is automatically called when unmarshalling a []byte into a CredentialManifest.
-func (cm *CredentialManifest) Validate() error {
+func (cm *CredentialManifest) validate() error {
 	if cm.Issuer.ID == "" {
 		return errors.New("issuer ID missing")
 	}
@@ -192,126 +151,6 @@ func (cm *CredentialManifest) Validate() error {
 	}
 
 	return nil
-}
-
-// ResolveFulfillment resolves given credential fulfillment and returns results.
-// Currently supporting only 'ldp_vc' format of fulfillment credentials.
-func (cm *CredentialManifest) ResolveFulfillment(fulfillment *verifiable.Presentation) ([]*ResolvedDescriptor, error) { //nolint:funlen,gocyclo,lll
-	var results []*ResolvedDescriptor
-
-	credentialFulfillmentMap, ok := lookupMap(fulfillment.CustomFields, "credential_fulfillment")
-	if !ok {
-		return nil, errors.New("invalid credential fulfillment")
-	}
-
-	if manifestID, k := credentialFulfillmentMap["manifest_id"]; !k || cm.ID != manifestID {
-		return nil, errors.New("credential fulfillment not matching")
-	}
-
-	descriptorMaps, ok := lookupArray(credentialFulfillmentMap, "descriptor_map")
-	if !ok {
-		return nil, errors.New("invalid descriptor map")
-	}
-
-	if len(descriptorMaps) == 0 {
-		return results, nil
-	}
-
-	outputDescriptors := mapDescriptors(cm)
-
-	builder := gval.Full(jsonpath.PlaceholderExtension())
-
-	vpBits, err := fulfillment.MarshalJSON()
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal vp: %w", err)
-	}
-
-	typelessVP := interface{}(nil)
-
-	err = json.Unmarshal(vpBits, &typelessVP)
-	if err != nil {
-		return nil, fmt.Errorf("failed to unmarshal vp: %w", err)
-	}
-
-	for _, descriptorMap := range descriptorMaps {
-		descriptor, ok := descriptorMap.(map[string]interface{})
-		if !ok {
-			return nil, errors.New("invalid descriptor format")
-		}
-
-		id, ok := lookupString(descriptor, "id")
-		if !ok {
-			return nil, errors.New("invalid descriptor ID")
-		}
-
-		outputDescriptor, ok := outputDescriptors[id]
-		if !ok {
-			return nil, errors.New("unable to find matching output descriptor from manifest")
-		}
-
-		if format, k := lookupString(descriptor, "format"); !k || format != "ldp_vc" {
-			// currently, only ldp_vc format is supported
-			continue
-		}
-
-		path, ok := lookupString(descriptor, "path")
-		if !ok {
-			return nil, fmt.Errorf("invalid credential path in descriptor '%s'", id)
-		}
-
-		credential, err := selectVCByPath(builder, typelessVP, path)
-		if err != nil {
-			return nil, fmt.Errorf("failed to select vc from descriptor: %w", err)
-		}
-
-		resolved, err := resolveOutputDescriptor(outputDescriptor, credential)
-		if err != nil {
-			return nil, fmt.Errorf("failed to resolve credential by descriptor: %w", err)
-		}
-
-		results = append(results, resolved)
-	}
-
-	return results, nil
-}
-
-// ResolveCredential resolved given credential and returns results.
-func (cm *CredentialManifest) ResolveCredential(descriptorID string, credential CredentialToResolveOption) (*ResolvedDescriptor, error) { //nolint:lll
-	opts := &resolveCredOpts{}
-
-	if credential != nil {
-		credential(opts)
-	}
-
-	var err error
-
-	var vcmap map[string]interface{}
-
-	switch {
-	case opts.credential != nil:
-		opts.rawCredential, err = opts.credential.MarshalJSON()
-		if err != nil {
-			return nil, err
-		}
-
-		fallthrough
-	case len(opts.rawCredential) > 0:
-		err = json.Unmarshal(opts.rawCredential, &vcmap)
-		if err != nil {
-			return nil, err
-		}
-	default:
-		return nil, errors.New("credential to resolve is not provided")
-	}
-
-	// find matching descriptor and resolve.
-	for _, descriptor := range cm.OutputDescriptors {
-		if descriptor.ID == descriptorID {
-			return resolveOutputDescriptor(descriptor, vcmap)
-		}
-	}
-
-	return nil, errors.New("unable to find matching descriptor")
 }
 
 // ValidateOutputDescriptors checks the given slice of OutputDescriptors to ensure that they are valid (per the spec)
@@ -351,107 +190,6 @@ func (cm *CredentialManifest) hasFormat() bool {
 	return hasAnyAlgorithmsOrProofTypes(cm.Format)
 }
 
-func resolveOutputDescriptor(outputDescriptor *OutputDescriptor,
-	vc map[string]interface{}) (*ResolvedDescriptor, error) {
-	var resolved ResolvedDescriptor
-
-	staticDisplayMappings, err := resolveStaticDisplayMappingObjects(outputDescriptor, vc)
-	if err != nil {
-		return nil, err
-	}
-
-	resolved.DescriptorID = outputDescriptor.ID
-	resolved.Title = staticDisplayMappings.title
-	resolved.Subtitle = staticDisplayMappings.subtitle
-	resolved.Description = staticDisplayMappings.description
-	resolved.Styles = outputDescriptor.Styles
-
-	resolved.Properties, err =
-		resolveDescriptorProperties(outputDescriptor.Display.Properties, vc)
-	if err != nil {
-		return nil, fmt.Errorf("failed to resolve properties: %w", err)
-	}
-
-	return &resolved, nil
-}
-
-func resolveStaticDisplayMappingObjects(outputDescriptor *OutputDescriptor,
-	vc map[string]interface{}) (staticDisplayMappingObjects, error) {
-	title, err := resolveDisplayMappingObject(&outputDescriptor.Display.Title, vc)
-	if err != nil {
-		return staticDisplayMappingObjects{}, fmt.Errorf("failed to resolve title display mapping object: %w", err)
-	}
-
-	subtitle, err := resolveDisplayMappingObject(&outputDescriptor.Display.Subtitle, vc)
-	if err != nil {
-		return staticDisplayMappingObjects{}, fmt.Errorf("failed to resolve subtitle display mapping object: %w", err)
-	}
-
-	description, err := resolveDisplayMappingObject(&outputDescriptor.Display.Description, vc)
-	if err != nil {
-		return staticDisplayMappingObjects{}, fmt.Errorf("failed to resolve description display mapping object: %w", err)
-	}
-
-	return staticDisplayMappingObjects{
-		title:       fmt.Sprintf("%v", title),
-		subtitle:    fmt.Sprintf("%v", subtitle),
-		description: fmt.Sprintf("%v", description),
-	}, nil
-}
-
-func resolveDescriptorProperties(properties []LabeledDisplayMappingObject,
-	vc map[string]interface{}) ([]*ResolvedProperty, error) {
-	var resolvedProperties []*ResolvedProperty
-
-	for _, property := range properties {
-		var err error
-
-		value, err := resolveDisplayMappingObject(&property.DisplayMappingObject, vc)
-		if err != nil {
-			return nil, fmt.Errorf("failed to resolve the display mapping object for the property with label '%s': %w", property.Label, err) // nolint:lll
-		}
-
-		resolvedProperties = append(resolvedProperties, &ResolvedProperty{
-			Label:  property.Label,
-			Schema: property.Schema,
-			Value:  value,
-		})
-	}
-
-	return resolvedProperties, nil
-}
-
-func resolveDisplayMappingObject(displayMappingObject *DisplayMappingObject,
-	vc map[string]interface{}) (interface{}, error) {
-	if len(displayMappingObject.Paths) > 0 {
-		resolvedValue, err := resolveJSONPathsUsingVC(displayMappingObject.Paths, displayMappingObject.Fallback, vc)
-		if err != nil {
-			return "", err
-		}
-
-		return resolvedValue, nil
-	}
-
-	return displayMappingObject.Fallback, nil
-}
-
-func resolveJSONPathsUsingVC(paths []string, fallback string, vc map[string]interface{}) (interface{}, error) {
-	for _, path := range paths {
-		resolvedValue, err := jsonpath.Get(path, vc)
-		if err != nil {
-			if strings.HasPrefix(err.Error(), "unknown key") {
-				continue
-			}
-
-			return nil, err
-		}
-
-		return resolvedValue, nil
-	}
-
-	return fallback, nil
-}
-
 func (cm *CredentialManifest) standardUnmarshal(data []byte) error {
 	// The type alias below is used as to allow the standard json.Unmarshal to be called within a custom unmarshal
 	// function without causing infinite recursion. See https://stackoverflow.com/a/43178272 for more information.
@@ -484,11 +222,11 @@ func validateOutputDescriptor(outputDescriptor *OutputDescriptor, outputDescript
 			outputDescriptorIndex, err)
 	}
 
-	for propertyIndex, property := range outputDescriptor.Display.Properties {
-		err := validateDisplayMappingObject(&property.DisplayMappingObject)
+	for i := range outputDescriptor.Display.Properties {
+		err := validateDisplayMappingObject(&outputDescriptor.Display.Properties[i].DisplayMappingObject)
 		if err != nil {
 			return fmt.Errorf("display property at index %d for output descriptor at index %d is invalid: %w",
-				outputDescriptorIndex, propertyIndex, err)
+				outputDescriptorIndex, i, err)
 		}
 	}
 
@@ -505,6 +243,8 @@ func validateDisplayMappingObject(displayMappingObject *DisplayMappingObject) er
 		}
 
 		return validateSchema(displayMappingObject)
+	} else if displayMappingObject.Text == "" {
+		return fmt.Errorf(`display mapping object must contain either a paths or a text property`)
 	}
 
 	return nil
@@ -534,7 +274,7 @@ func validateSchema(displayMappingObject *DisplayMappingObject) error {
 func schemaFormatIsValid(format string) bool {
 	validFormats := []string{
 		"", "date-time", "time", "date", "email", "idn-email", "hostname", "idn-hostname",
-		"ipv4", "ipv6", "uri", "uri-reference", "iri", "iri-reference", "image/png",
+		"ipv4", "ipv6", "uri", "uri-reference", "iri", "iri-reference",
 	}
 
 	var isValidFormat bool
@@ -547,39 +287,6 @@ func schemaFormatIsValid(format string) bool {
 	}
 
 	return isValidFormat
-}
-
-func lookupString(model map[string]interface{}, key string) (string, bool) {
-	raw, ok := model[key]
-	if !ok {
-		return "", false
-	}
-
-	val, ok := raw.(string)
-
-	return val, ok
-}
-
-func lookupMap(model map[string]interface{}, key string) (map[string]interface{}, bool) {
-	raw, ok := model[key]
-	if !ok {
-		return nil, false
-	}
-
-	val, ok := raw.(map[string]interface{})
-
-	return val, ok
-}
-
-func lookupArray(model map[string]interface{}, key string) ([]interface{}, bool) {
-	raw, ok := model[key]
-	if !ok {
-		return nil, false
-	}
-
-	val, ok := raw.([]interface{})
-
-	return val, ok
 }
 
 func mapDescriptors(manifest *CredentialManifest) map[string]*OutputDescriptor {
